warning fixes for sim directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23643 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/sim/V2/runtime/simv2_NameList.C
@@ -157,7 +157,7 @@ simv2_NameList_getName(visit_handle h, int i, char **val)
     }
     int retval = VISIT_ERROR;
     VisIt_NameList *obj = GetObject(h, "simv2_NameList_getName");
-    if(obj != NULL && i >= 0 && i < obj->names.size())
+    if(obj != NULL && i >= 0 && i < (int)obj->names.size())
     {
         *val = (char *)malloc(obj->names[i].size() + 1);
         strcpy(*val, obj->names[i].c_str());
